Revert "Feature/remove defaultimpl"
