+++ Instantiate/Extensions.swift
@@ -1,0 +1,19 @@
+//
+//  Extensions.swift
+//  Instantiate
+//
+//  Created by tarunon on 2016/12/04.
+//  Copyright © 2016年 tarunon. All rights reserved.
+//
+
+import Foundation
+
+extension NSObject {
+    static var className: String {
+        return NSStringFromClass(self).components(separatedBy: ".").last!
+    }
+    
+    static var bundle: Bundle {
+        return Bundle(for: self)
+    }
+}
