+++ Instantiate/Identifier.swift
@@ -16,6 +16,12 @@ public struct Identifier {
     }
 }
 
+extension Identifier {
+    public init<O: NSObject>(type: O.Type) {
+        self.rawValue = O.className
+    }
+}
+
 extension Identifier: ExpressibleByStringLiteral {
     public typealias UnicodeScalarLiteralType = String
     public typealias ExtendedGraphemeClusterLiteralType = String
@@ -29,8 +35,4 @@ extension Identifier: ExpressibleByStringLiteral {
     public init(unicodeScalarLiteral value: String) {
         self.rawValue = value
     }
-}
-
-public func +(lhs: Identifier, rhs: Identifier) -> Identifier {
-    return Identifier(rawValue: lhs.rawValue + rhs.rawValue)
 }
