--- InstantiateDefaultMethods/InstantiateDefaultMethods.h
@@ -1,19 +1,0 @@
-//
-//  InstantiateDefaultMethods.h
-//  InstantiateDefaultMethods
-//
-//  Created by tarunon on 2016/12/05.
-//  Copyright © 2016年 tarunon. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for InstantiateDefaultMethods.
-FOUNDATION_EXPORT double InstantiateDefaultMethodsVersionNumber;
-
-//! Project version string for InstantiateDefaultMethods.
-FOUNDATION_EXPORT const unsigned char InstantiateDefaultMethodsVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <InstantiateDefaultMethods/PublicHeader.h>
-
-
